Add files via upload
This code updates many ads such as washing machines , folding machines , ironing boards and other stuff every X minutes from 07:00:00 (in the morning) until 23:55:00 (night) ,where X is computed from a function in order to update 200 ads equivalent between that period.
I run this code for ever in a raspberry pi Zero (Ubuntu server ~ without graphical environment) in order to be faster.
In the beginning of the day, informs you with an email that the updates started and at 23:55:00 it sends you an email informs you the total updates , number of updates for
each ad.All these information hold in 7 files. At the end of the day (23:55:00) ,all files  being reset and finally all this work runs again the other day.